% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{chicago_pal}
\alias{chicago_pal}
\title{Chicago color palettes}
\usage{
chicago_pal(palette = "brand", discrete = TRUE)
}
\arguments{
\item{palette}{Chicago color palette. One of 'brand', 'base', 'diverging', 'sequential'}

\item{discrete}{Whether you want a discrete palette or not}
}
\description{
Color palettes drawn from Chicago's branding guidelines
}
\examples{
scales::show_col(chicago_pal()(7))
scales::show_col(chicago_pal('diverging')(7))
scales::show_col(chicago_pal(discrete = FALSE)(0.1*0:10))
scales::show_col(chicago_pal('diverging', discrete = FALSE)(0.1*0:10))
}
\concept{chicago colors}
