% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggchicago-data.R
\name{get_colors}
\alias{get_colors}
\alias{get_available_colors}
\alias{get_palette}
\alias{get_available_palettes}
\title{Functions for accessing colors and color palettes}
\usage{
get_colors(colors)

get_available_colors()

get_palette(palette)

get_available_palettes()
}
\arguments{
\item{colors}{Color names to grab}

\item{palette}{Color palette name}
}
\description{
Functions for accessing colors and color palettes
}
\examples{
get_colors(c('rusty-red', 'atomic-green'))
get_available_colors()
get_palette('brand')
get_available_palettes()
}
